modify data and model architect
