+++ src/llamafactory/data/loader.py
@@ -21,7 +21,7 @@ from datasets import Dataset, load_dataset, load_from_disk
 from ..extras import logging
 from ..extras.constants import FILEEXT2TYPE
 from ..extras.misc import check_version, has_tokenized_data
-from .converter import align_dataset
+from .data_converter.converter import align_dataset
 from .data_utils import get_dataset_module, merge_dataset, read_cloud_json, split_dataset
 from .parser import get_dataset_list
 from .processor import (
@@ -30,7 +30,7 @@ from .processor import (
     PairwiseDatasetProcessor,
     PretrainDatasetProcessor,
     SupervisedDatasetProcessor,
-    TargetingDatasetProcessor,
+    NluHeadDatasetProcessor,
     NoTemplateDatasetProcessor,
     UnsupervisedDatasetProcessor,
 )
@@ -219,8 +219,8 @@ def _get_dataset_processor(
         else:
             if data_args.dynamic:
                 dataset_processor_class = NoTemplateDatasetProcessor
-            elif data_args.targeting:
-                dataset_processor_class = TargetingDatasetProcessor
+            elif data_args.target_nlu:
+                dataset_processor_class = NluHeadDatasetProcessor
             else:
                 dataset_processor_class = SupervisedDatasetProcessor
