+++ src/llamafactory/model/loader.py
@@ -36,7 +36,7 @@ from .model_utils.misc import register_autoclass
 from .model_utils.mod import convert_pretrained_model_to_mod, load_mod_pretrained_model
 from .model_utils.unsloth import load_unsloth_pretrained_model
 from .model_utils.valuehead import load_valuehead_params
-from .model_utils.task_plugin_model import QwenWithTaskPlugin
+from .model_utils.nlu_model.task_plugin_model import QwenWithTaskPlugin
 from .patcher import patch_config, patch_model, patch_processor, patch_tokenizer, patch_valuehead_model
